return qulonglong
svn path=/branches/work/kdab-post-4.0/kdepim/; revision=738628

+++ kleopatra/uiserver/detail_p.h
@@ -88,7 +88,7 @@ namespace _detail {
 
     //returns an integer representation of the assuan_fd_t,
     //suitable for debug output
-    static inline int assuanFD2int( assuan_fd_t fd )
+    static inline qulonglong assuanFD2int( assuan_fd_t fd )
     {
 #ifdef Q_OS_WIN32
         return reinterpret_cast<qulonglong>( fd );
